requirements.txt
pillow == 6.2.1

twine >= 1.8.0
setuptools >= 38.6.0
wheel

pytest
pyside2
